validator check active as worker
